Api naming

+++ VSC.Toolsy.Server/Controllers/AdminController.cs
@@ -2,6 +2,7 @@
 using VSC.Toolsy.Common.DTOs.Responses;
 using VSC.Toolsy.Common.Enums;
 using VSC.Toolsy.Common.Interfaces;
+using VSC.Toolsy.Common.Models.CoreEntites;
 
 namespace VSC.Toolsy.Server.Controllers
 {
@@ -18,9 +19,17 @@ namespace VSC.Toolsy.Server.Controllers
         }
 
         [HttpGet("users")]
+        [ProducesResponseType(typeof(ApiResponseDto<Profile>), StatusCodes.Status200OK)] // OK - 200 status code 
+        [ProducesResponseType(typeof(ApiResponseDto<string>), StatusCodes.Status500InternalServerError)] // Internal Server Error - 500 status code 
+        [ProducesResponseType(typeof(ApiResponseDto<string>), StatusCodes.Status404NotFound)] // Not Found - 404 status code
         public async Task<IActionResult> GetAllUsers()
-        { 
+        {
             List<AdminUserDto> users = await _adminService.GetAllUsersForAdminAsync();
+
+            if (users == null)
+            {
+                return NotFound(ApiResponseDto<string>.FailureResponse("Users not found."));
+            }
             return Ok(ApiResponseDto<List<AdminUserDto>>.SuccessResponse(
                 users,
                 "Successfully fetched all users for the admin."
@@ -28,32 +37,55 @@ namespace VSC.Toolsy.Server.Controllers
         }
 
         [HttpGet("users/verified")]
+        [ProducesResponseType(typeof(ApiResponseDto<Profile>), StatusCodes.Status200OK)] // OK - 200 status code 
+        [ProducesResponseType(typeof(ApiResponseDto<string>), StatusCodes.Status500InternalServerError)] // Internal Server Error - 500 status code 
+        [ProducesResponseType(typeof(ApiResponseDto<string>), StatusCodes.Status404NotFound)] // Not Found - 404 status code
         public async Task<IActionResult> GetAllVerifiedUsers()
         {
             List<AdminUserDto> users = (await _adminService.GetAllUsersForAdminAsync())
                 .Where(u => u.VerificationStatus.Equals(VerificationStatus.Verified))
                 .ToList();
 
+            if (users == null || users.Count == 0)
+            {
+                return NotFound(ApiResponseDto<string>.FailureResponse("No verified users found."));
+            }
+
             return Ok(ApiResponseDto<List<AdminUserDto>>.SuccessResponse(
                 users,
                 "Successfully fetched all users for the admin."
             ));
+
+
         }
 
-        [HttpGet("users/unVerified")]
+        [HttpGet("users/unverified")]
+        [ProducesResponseType(typeof(ApiResponseDto<Profile>), StatusCodes.Status200OK)] // OK - 200 status code 
+        [ProducesResponseType(typeof(ApiResponseDto<string>), StatusCodes.Status500InternalServerError)] // Internal Server Error - 500 status code 
+        [ProducesResponseType(typeof(ApiResponseDto<string>), StatusCodes.Status404NotFound)] // Not Found - 404 status code
         public async Task<IActionResult> GetAllUnVerifiedUsers()
         {
             List<AdminUserDto> users = (await _adminService.GetAllUsersForAdminAsync())
                 .Where(u => u.VerificationStatus.Equals(VerificationStatus.Pending))
                 .ToList();
 
+            if (users == null || users.Count == 0)
+            {
+                return NotFound(ApiResponseDto<string>.FailureResponse("No Unverified users found."));
+            }
+
             return Ok(ApiResponseDto<List<AdminUserDto>>.SuccessResponse(
                 users,
                 "Successfully fetched all users for the admin."
             ));
+
+
         }
 
         [HttpGet("users/deleted")]
+        [ProducesResponseType(typeof(ApiResponseDto<Profile>), StatusCodes.Status200OK)] // OK - 200 status code
+        [ProducesResponseType(typeof(ApiResponseDto<string>), StatusCodes.Status500InternalServerError)] // Internal Server Error - 500 status code
+        [ProducesResponseType(typeof(ApiResponseDto<string>), StatusCodes.Status404NotFound)] // Not Found - 404 status code
         public async Task<IActionResult> GetAllDeletedUsers()
         {
 
@@ -61,39 +93,91 @@ namespace VSC.Toolsy.Server.Controllers
                 .Where(u => u.IsDeleted)
                 .ToList();
 
+            if (users == null || users.Count == 0)
+            {
+                return NotFound(ApiResponseDto<string>.FailureResponse("No Deleted users found."));
+            }
+
             return Ok(ApiResponseDto<List<AdminUserDto>>.SuccessResponse(
                 users,
                 "Successfully fetched all users for the admin."
             ));
 
+
         }
 
         [HttpPut("profiles/approve")]
+        [ProducesResponseType(typeof(ApiResponseDto<Profile>), StatusCodes.Status200OK)] // OK - 200 status code
+        [ProducesResponseType(typeof(ApiResponseDto<string>), StatusCodes.Status500InternalServerError)] // Internal Server Error - 500 status code
+        [ProducesResponseType(typeof(ApiResponseDto<string>), StatusCodes.Status404NotFound)] // Not Found - 404 status code
+        [ProducesResponseType(typeof(ValidationProblemDetails), StatusCodes.Status400BadRequest)] // Bad Request - 400 status code
         public async Task<IActionResult> ApproveProfileAccount(string email)
         {
 
+            if (email == null)
+            {
+                return BadRequest(ApiResponseDto<string>.FailureResponse("Email cannot be empty"));
+            }
+
             bool result = await _adminService.ApproveProfileAccountAsync(email);
+
+            if (!result)
+            {
+                return NotFound(ApiResponseDto<string>.FailureResponse("User not found."));
+            }
 
             return Ok(ApiResponseDto<bool>.SuccessResponse(result, "Approve Account"));
 
+
+
         }
 
-        [HttpPut("users/deleteByEmail")]
+        [HttpPut("user/delete-by-email")]
+        [ProducesResponseType(typeof(ApiResponseDto<Profile>), StatusCodes.Status200OK)] // OK - 200 status code
+        [ProducesResponseType(typeof(ApiResponseDto<string>), StatusCodes.Status500InternalServerError)] // Internal Server Error - 500 status code
+        [ProducesResponseType(typeof(ApiResponseDto<string>), StatusCodes.Status404NotFound)] // Not Found - 404 status code
+        [ProducesResponseType(typeof(ValidationProblemDetails), StatusCodes.Status400BadRequest)] // Bad Request - 400 status code
         public async Task<IActionResult> DeleteUserAccount(string email)
         {
+            if (email == null)
+            {
+                return BadRequest(ApiResponseDto<string>.FailureResponse("Email cannot be empty"));
+            }
 
             bool result = await _adminService.DeleteUserAccountAsync(email);
 
+            if (!result)
+            {
+                return NotFound(ApiResponseDto<string>.FailureResponse("User not found or could not be deleted."));
+            }
+
             return Ok(ApiResponseDto<bool>.SuccessResponse(result, "Approve Account"));
+
         }
 
-        [HttpGet("users/getByEmail")]
-        public async Task<IActionResult> GetUserById(string email)
+        [HttpGet("user/email")]
+        [ProducesResponseType(typeof(ApiResponseDto<Profile>), StatusCodes.Status200OK)] // OK - 200 status code
+        [ProducesResponseType(typeof(ApiResponseDto<string>), StatusCodes.Status500InternalServerError)] // Internal Server Error - 500 status code
+        [ProducesResponseType(typeof(ValidationProblemDetails), StatusCodes.Status400BadRequest)] // Bad Request - 400 status code
+        [ProducesResponseType(typeof(ApiResponseDto<string>), StatusCodes.Status404NotFound)] // Not Found - 404 status code
+        public async Task<IActionResult> GetUserByEmail(string email)
         {
+            if (email == null)
+            {
+                return BadRequest(ApiResponseDto<string>.FailureResponse("Email cannot be empty"));
+            }
 
             AdminUserDto adminUserDto = await _adminService.GetUserByEmailAsync(email);
 
+
+            if (adminUserDto == null)
+            {
+                return NotFound(ApiResponseDto<string>.FailureResponse("User not found."));
+            }
+
+
             return Ok(ApiResponseDto<AdminUserDto>.SuccessResponse(adminUserDto, "GetUserByEmailAsync"));
+
         }
     }
 }
